Fix a critical bug with the prefix not being fetched correctly
The prefix was being fetched for every channel once, every time any channel updated its prefix, and there was only one prefix variable. That meant that any update to the prefix would cause that update to propagate for every channel. Fix this by storing all prefixes in a dictionary indexed by channel IDs.

Also adds the eval and setvar functions, and changes the way scv is exported to match the way cmd is exported.