Add ability to send messages greater than the char limit
Add function safeSendMsg, which detects if a message will break the char limit and if so sends it in chunks. The constant MY_CHAR_LIMIT is there to prevent abuse; I think 10000 will do for now. Any message greater than this will not be sent (as opposed to 2000 by sending normally).