Add param count checking, cleanup code
Add param count checking to cmdExe. Now, if a the user attempts to execute a command with too few arguments, the bot will not allow the command to proceed to execution and fail there with undefined behaviour, but rather inform the user of their mistake. Add the `defaults` field to commands.json. This field specifies how many default parameters a command has; the minimum amount of arguments is therefore the amount of parameters minus the amount of default parameters. Note: it is not needed to check whether the user has supplied *more* arguments than required, since this is impossible, as any excess arguments are squished into the last argument by default.