Fix bug where hangman would give multiple messages upon initialisation of a game
The issue was that every time ~hm start was run, an event listener waiting for PMs from the user was created; however, when it timed out (on purpose) or when a satisfactory PM with settings for the game was received, the event listener was not disabled. As a result, event listeners accumulated upon each call of ~hm start, which caused the issue with multiple responses to the PM. I would imagine that, on a greater scale, this would have also caused a significant drop in performance and an increase in latency, due to the massive increase of redundant event handlers.

Fix the bug by saving event handlers and handler timeouts in a dictionary, with keys being the channel IDs (since each channel can only have one hangman at a time). Use client.removeListener to remove any redundant listeners upon ~hm start, and upon initialising the game.